Moved more stuff to core, do not release demo app.
