+++ reactive-spring-session-core/src/main/java/net/andreaskluth/session/core/serializer/DeserializationException.java
@@ -1,0 +1,13 @@
+package net.andreaskluth.session.core.serializer;
+
+import net.andreaskluth.session.core.ReactiveSessionException;
+
+public class DeserializationException extends ReactiveSessionException {
+
+  private static final long serialVersionUID = 2260800020294378692L;
+
+  public DeserializationException(Throwable throwable) {
+
+    super(throwable);
+  }
+}
