+++ reactive-spring-session-core/src/main/java/net/andreaskluth/session/core/serializer/SerializationStrategy.java
@@ -1,4 +1,4 @@
-package net.andreaskluth.session.postgres.serializer;
+package net.andreaskluth.session.core.serializer;
 
 import java.util.Map;
