+++ reactive-spring-session-core/src/test/java/net/andreaskluth/session/core/serializer/JdkSerializationStrategyTest.java
@@ -1,4 +1,4 @@
-package net.andreaskluth.session.postgres.serializer;
+package net.andreaskluth.session.core.serializer;
 
 class JdkSerializationStrategyTest extends TestSerializationStrategyBase {
