+++ reactive-spring-session-postgres/src/main/java/net/andreaskluth/session/postgres/ReactivePostgresSessionConfiguration.java
@@ -8,8 +8,8 @@ import io.vertx.sqlclient.Pool;
 import io.vertx.sqlclient.PoolOptions;
 import java.time.Clock;
 import java.util.Optional;
-import net.andreaskluth.session.postgres.serializer.JdkSerializationStrategy;
-import net.andreaskluth.session.postgres.serializer.SerializationStrategy;
+import net.andreaskluth.session.core.serializer.JdkSerializationStrategy;
+import net.andreaskluth.session.core.serializer.SerializationStrategy;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
